chore: add GitHub actions and remove Package wapproj (#13)
Adds fmt, test and build jobs using dotnet.

The fmt job currently doesn't match what Rider does, but I think that
will be difficult to pull off without using a different formatting tool
which I don't want to investigate for now.

Removes the `Package` project in favor of making the `App` project
include the necessary stuff for packaging itself. We weren't getting any
benefit from it being in a separate package yet and it was not building
correctly in dotnet (only in VS) when they were separated.

Closes #7