feat: adds mutagen daemon controller (#58)
fixes coder/internal#380

Adds MutagenController service, which manages the lifecycle of the `mutagen` daemon and exposes methods to query and modify sync sessions.

Sync sessions themselves are a placeholder which will need to be filled out and plumbed thru to the actual API. This just handles daemon lifecycle for now.