fix: fix Downloader to dispose tempFile and use synchronous IO  (#81)
Fixes https://github.com/coder/internal/issues/598

There is a possible race where if the cancellation token is expired, `Download()` never gets called and the tempFile is never disposed of (at least until GC). We also switch to synchronous IO so that a pending overlapped write won't block the deletion.

These issues can cause races in our tests when we try to clean up the directory.